requirements.txt
streamlit==1.37.1
pandas==2.2.2
openpyxl==3.1.5
python-docx==1.1.2
